requirements.txt
annotated-types==0.7.0
dj-database-url==2.1.0
asgiref==3.8.1
Brotli==1.1.0
certifi==2025.4.26
cffi==1.17.1
chardet==5.2.0
charset-normalizer==3.4.2
colorama==0.4.6
cssselect2==0.8.0
Django==5.2.1
django-qr-code==4.2.0
django-widget-tweaks==1.5.0
fonttools==4.58.0
idna==3.10
numpy==2.3.0
pandas==2.3.0
pillow==11.2.1
psycopg2-binary==2.9.10
pycparser==2.22
pydantic==2.11.4
pydantic_core==2.33.2
pydyf==0.11.0
pyphen==0.17.2
python-barcode==0.15.1
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
pytz==2025.2
qrcode==8.2
reportlab==4.4.1
requests==2.32.4
segno==1.6.6
setuptools==80.7.1
six==1.17.0
sqlparse==0.5.3
tinycss2==1.4.0
tinyhtml5==2.0.0
typing-inspection==0.4.0
typing_extensions==4.13.2
tzdata==2025.2
urllib3==2.4.0
weasyprint==65.1
webencodings==0.5.1
zopfli==0.2.3.post1
# django-tenants removed (moving to single tenant)

# Dependencias adicionales para Docker
# Servidor de alto rendimiento (ASGI) y WebSockets
granian==1.6.0
channels==4.0.0
daphne==4.0.0
channels-redis==4.2.0
redis==5.0.1
django-redis==5.4.0
celery==5.3.4
whitenoise==6.6.0
dj-database-url==2.1.0
djangorestframework
pyserial
python-escpos
django-crispy-forms
crispy-bootstrap5
pyusb
zeep==4.3.2
cryptography
